Fix minor errors

+++ admin/delete.php
@@ -1,57 +1,53 @@
 <?php 
-//include ('../adminHeader.php');
-include('../connDB.php');
+session_start();
+if(isset($_SESSION['userID']) && $_SESSION['regisType'] == "A"){
+    include('../connDB.php');
 
 
 
-//<a href="../view.php?regisType=U"><img src='../../button/back.png' width="50" height="50"></a> 
 
+    if(isset($_REQUEST["ID"]) && isset($_REQUEST['type'])){
 
-if(isset($_REQUEST["ID"]) && isset($_REQUEST['type'])){
+     $ID = $_REQUEST["ID"];
+     $type = $_REQUEST["type"];
 
-   $ID = $_REQUEST["ID"];
-   $type = $_REQUEST["type"];
+     switch ($type) {
+        case 'U':
+        $query = "DELETE FROM users WHERE ID = $ID";
+        break;
+        case 'P':
+        $query = "DELETE FROM patients WHERE ID = $ID";
+        break;
+        case 'C':
+        $query = "DELETE FROM clients WHERE ID = $ID";
+        break;
+        default:
 
-   switch ($type) {
-    case 'U':
-    $query = "DELETE FROM users WHERE ID = $ID";
-    break;
-    case 'P':
-    $query = "DELETE FROM patients WHERE ID = $ID";
-    break;
-    case 'C':
-    $query = "DELETE FROM clients WHERE ID = $ID";
-    break;
-    default:
+        break;
+    }
 
-    break;
-}
+    if(isset($query)){
+        $result = mysqli_query($con,$query);
 
-if(isset($query)){
-    $result = mysqli_query($con,$query);
-
-    $return = 'Location: view.php?type=' . $type;
-    if ($result) {
+        $return = 'Location: view.php?type=' . $type;
+        if ($result) {
             header($return);
         }else{
             echo 'failed to delete';
         }   
 
-}else{
-        header("HTTP/1.0 404 Not Found");
-    echo "<h1>404 Not Found</h1>";
-    echo "<p>The page that you have requested could not be found.</p>";
+    }else{
+        echo "Query is NULL";
 
-
-    exit();
-}
+        exit();
+    }
 
 
 
 
 }
 else{
-            header("HTTP/1.0 404 Not Found");
+    header("HTTP/1.0 404 Not Found");
     echo "<h1>404 Not Found</h1>";
     echo "<p>The page that you have requested could not be found.</p>";
 
@@ -59,4 +55,14 @@ else{
     exit();
 }
 
+}else{
+    header("HTTP/1.0 404 Not Found");
+    echo "<h1>404 Not Found</h1>";
+    echo "<p>The page that you have requested could not be found.</p>";
+
+    echo "<a href= '../login.php'>";
+    echo "Please Login to open this page.";
+    echo "</a>";
+    exit();
+}
 ?>
